Log file parameters have been changed a bit for BBs
