--- src/.gitignore
@@ -1,18 +1,0 @@
-/githash.h
-
-# The same ignores apply to all compilation folders
-/*/*.exe
-/*/*.dll
-/*/adda*
-/*/*.o
-/*/*.d
-/*/.*opts
-/*/*stackdump
-/*/ExpCount
-/*/tables/
-/*/run*/
-/*/test*/
-/*/*.dat
-/*/*.geom
-/*/out
-/*/*.clstr
