+++ src/githash.h
@@ -1,0 +1,1 @@
+#define GITHASH "b660860"
